xw1/requirements.txt
bottle==0.12.10
cassandra-driver==3.7.1
cql==1.4.0
cqlsh==5.0.3
futures==3.0.5
haversine==0.4.5
kafka-python==1.3.1
linecache2==1.0.0
mzgeohash==1.0.0
numpy==1.11.2
pandas==0.19.1
pygeohash==1.2.0
python-dateutil==2.6.0
pytz==2016.10
simplejson==3.10.0
six==1.10.0
thrift==0.9.3
traceback2==1.4.0
unittest2==1.1.0
